Update README.md
This commit updates the project's README file with a more detailed description of the problem, solution, and project goals. It clarifies that the project name is 'kafkactl' and that it installs two separate commands: 'kvm' for version management and 'kafka' for operations. It also notes that the project is inspired by tools like 'nvm' and 'gvm' to provide a familiar experience for developers.

Signed-off-by: Charan Ravela <56770996+CharanRavela@users.noreply.github.com>